Limit thumbnail download parallelism (#40)
* Limit thumbnail download parallelism

The number of concurrent thumbnail downloads is now limited to 5.
Introduced a separate dispatcher `thumbnailDispatcher` for this purpose.
Also, increased the parallelism for database operations via `dbDispatcher` to 8.

* Fetch records updated after the latest

The `updatedAt` parameter for fetching records from the server is adjusted to retrieve records updated strictly after the previous `updatedAt` value. This ensures that only newer records are fetched by subtracting 1 from the timestamp.