Refactor: Rename `status` field to `uiState` in `RecordModel`
The `status` field in the `RecordModel` data class, which was of type `RecordUiState`, has been renamed to `uiState` for clarity. The `@SerializedName` annotation has been updated from `"status"` to `"state"` to reflect this change in the JSON representation. A new field named `status` of type `RecordStatus` has been added with the `@SerializedName("status")` annotation.